Adding permissions for docs and enhancing the publish permission flow

+++ components/appmgt/org.wso2.carbon.appmgt.impl/src/main/java/org/wso2/carbon/appmgt/impl/UserAwareAPIProvider.java
@@ -44,19 +44,20 @@ public class UserAwareAPIProvider extends APIProviderImpl {
 
     @Override
     public void addWebApp(WebApp api) throws AppManagementException {
-//        checkCreatePermission();
+        checkCreatePermission();
         super.addWebApp(api);
     }
 
     @Override
     public void createNewAPIVersion(WebApp api, String newVersion) throws DuplicateAPIException,
                                                                           AppManagementException {
-//        checkCreatePermission();
+        checkCreatePermission();
         super.createNewAPIVersion(api, newVersion);
     }
 
     @Override
     public void updateAPI(WebApp api) throws AppManagementException {
+        checkWebappUpdatePermission();
 //        checkCreatePermission();
         super.updateAPI(api);
     }
@@ -64,19 +65,20 @@ public class UserAwareAPIProvider extends APIProviderImpl {
     @Override
     public void changeAPIStatus(WebApp api, APIStatus status, String userId,
                                 boolean updateGatewayConfig) throws AppManagementException {
-//        if(userId.contains("@")){
-//            checkPublishPermissionTenantMode(userId);
-//        }else {
-//        if(APIStatus.PUBLISHED.equals(status)) {
-//            checkPublishPermission();
-//        }
-//        }
+        if(userId.contains("@")){
+            checkPublishPermissionTenantMode(userId);
+        }else {
+        if(APIStatus.PUBLISHED.equals(status)) {
+            checkPublishPermission();
+        }
+        }
         super.changeAPIStatus(api, status, userId, updateGatewayConfig);
     }
 
     @Override
     public void addDocumentation(APIIdentifier apiId,
                                  Documentation documentation) throws AppManagementException {
+        checkDocumentAddPermission();
 //        checkCreatePermission();
         super.addDocumentation(apiId, documentation);
     }
@@ -84,6 +86,7 @@ public class UserAwareAPIProvider extends APIProviderImpl {
     @Override
     public void removeDocumentation(APIIdentifier apiId, String docName,
                                     String docType) throws AppManagementException {
+        checkDocumentRemovePermission();
 //        checkCreatePermission();
         super.removeDocumentation(apiId, docName, docType);
     }
@@ -91,6 +94,7 @@ public class UserAwareAPIProvider extends APIProviderImpl {
     @Override
     public void updateDocumentation(APIIdentifier apiId,
                                     Documentation documentation) throws AppManagementException {
+        checkDocumentEditPermission();
 //        checkCreatePermission();
         super.updateDocumentation(apiId, documentation);
     }
@@ -99,6 +103,7 @@ public class UserAwareAPIProvider extends APIProviderImpl {
     public void addDocumentationContent(APIIdentifier identifier, String documentationName,
                                         String text) throws AppManagementException {
 //        checkCreatePermission();
+        checkDocumentAddPermission();
         super.addDocumentationContent(identifier, documentationName, text);
     }
 
@@ -106,31 +111,53 @@ public class UserAwareAPIProvider extends APIProviderImpl {
     public void copyAllDocumentation(APIIdentifier apiId, String toVersion) throws
                                                                             AppManagementException {
 //        checkCreatePermission();
+        checkDocumentAddPermission();
         super.copyAllDocumentation(apiId, toVersion);
     }
 
-//    public void checkCreatePermission() throws AppManagementException {
-//        AppManagerUtil.checkPermission(username, AppMConstants.Permissions.API_CREATE);
-//    }
+    public void checkDocumentRemovePermission() throws AppManagementException {
+        AppManagerUtil.checkPermission(username, AppMConstants.Permissions.DOCUMENT_EDIT);
+    }
+
+    public void checkDocumentEditPermission() throws AppManagementException {
+        AppManagerUtil.checkPermission(username, AppMConstants.Permissions.DOCUMENT_EDIT);
+    }
+
+    public void checkDocumentAddPermission() throws AppManagementException {
+        AppManagerUtil.checkPermission(username, AppMConstants.Permissions.DOCUMENT_ADD);
+    }
+
+    public void checkCreatePermission() throws AppManagementException {
+        AppManagerUtil.checkPermission(username, AppMConstants.Permissions.WEB_APP_CREATE);
+    }
+
+    public void checkWebappUpdatePermission() throws AppManagementException {
+        AppManagerUtil.checkPermission(username, AppMConstants.Permissions.WEB_APP_UPDATE);
+    }
+
+    public void checkWebappDeletePermission() throws AppManagementException {
+        AppManagerUtil.checkPermission(username, AppMConstants.Permissions.WEB_APP_DELETE);
+    }
     
     public void checkManageTiersPermission() throws AppManagementException {
         AppManagerUtil.checkPermission(username, AppMConstants.Permissions.MANAGE_TIERS);
     }
 
-//    public void checkPublishPermission() throws AppManagementException {
-//        AppManagerUtil.checkPermission(username, AppMConstants.Permissions.API_PUBLISH);
-//    }
+    public void checkPublishPermission() throws AppManagementException {
+        AppManagerUtil.checkPermission(username, AppMConstants.Permissions.WEB_APP_PUBLISH);
+    }
 
     @Override
     public void generateEntitlementPolicies(APIIdentifier apiIdentifier) throws
                                                                          AppManagementException {
-//        checkCreatePermission();
+        checkCreatePermission();
         super.generateEntitlementPolicies(apiIdentifier);
     }
 
     @Override
     public void updateEntitlementPolicies(List<EntitlementPolicy> policies) throws
                                                                             AppManagementException {
+        checkWebappUpdatePermission();
 //        checkCreatePermission();
         super.updateEntitlementPolicies(policies);
     }
@@ -142,6 +169,7 @@ public class UserAwareAPIProvider extends APIProviderImpl {
     @Override
     public boolean deleteEntitlementPolicyPartial(int policyPartialId, String author) throws
                                                                                       AppManagementException {
+        checkWebappDeletePermission();
 //        checkCreatePermission();
         return super.deleteEntitlementPolicyPartial(policyPartialId, author);
     }
@@ -150,6 +178,7 @@ public class UserAwareAPIProvider extends APIProviderImpl {
     public boolean updateEntitlementPolicyPartial(int policyPartialId, String policyPartial,
                                                   String author, boolean isShared, String policyPartialDesc)
             throws AppManagementException {
+        checkWebappUpdatePermission();
 //        checkCreatePermission();
         return super.updateEntitlementPolicyPartial(policyPartialId, policyPartial, author, isShared,
                 policyPartialDesc);
@@ -160,7 +189,7 @@ public class UserAwareAPIProvider extends APIProviderImpl {
     public int saveEntitlementPolicyPartial(String policyPartialName, String policyPartial, boolean isSharedPartial,
                                             String policyAuthor, String policyPartialDesc)
             throws AppManagementException {
-//        checkCreatePermission();
+        checkCreatePermission();
         return super.saveEntitlementPolicyPartial(policyPartialName, policyPartial, isSharedPartial, policyAuthor,
                 policyPartialDesc);
     }
